requirements.txt
alabaster==0.7.13
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
attrs==23.1.0
Babel==2.12.1
backports.entry-points-selectable==1.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
card==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chart-studio==1.1.0
click==8.1.3
cmdstanpy==1.1.0
contourpy==1.0.7
convertdate==2.4.0
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
docutils==0.19
entrypoints==0.4
ephem==4.1.4
Faker==18.6.0
favicon==0.7.0
filelock==3.12.0
fonttools==4.39.3
frozendict==2.3.7
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
hijri-converter==2.3.1
holidays==0.23
htbuilder==0.6.1
html5lib==1.1
httpcore==0.17.0
httpx==0.24.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
multitasking==0.0.11
newsapi-python==0.2.7
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathlib==1.0.1
Pillow==9.5.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
polygon==1.1.0
prophet==1.1.2
protobuf==3.20.3
pyarrow==11.0.0
pycountry==22.3.5
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==9.11
PyMeeus==0.5.12
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
retrying==1.3.4
rich==13.3.5
rows==0.4.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
Sphinx==7.0.0
sphinx-autodoc-typehints==1.23.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
st-annotated-text==4.0.0
streamlit==1.22.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-javascript==0.1.5
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unicodecsv==0.14.1
urllib3==2.0.1
validators==0.20.0
virtualenv==20.23.0
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
yarg==0.1.9
yfinance==0.2.18
zipp==3.15.0
